requirements.txt
beautifulsoup4==4.11.1
firebase_admin==5.3.0
pandas==1.4.4
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.28.1
typing_extensions==4.3.0
web3==5.30.0
